+++ table/iterator_wrapper.h
@@ -19,19 +19,21 @@ namespace rocksdb {
 // the valid() and key() results for an underlying iterator.
 // This can help avoid virtual function calls and also gives better
 // cache locality.
-class IteratorWrapper {
+template <class TValue = Slice>
+class IteratorWrapperBase {
  public:
-  IteratorWrapper() : iter_(nullptr), valid_(false) {}
-  explicit IteratorWrapper(InternalIterator* _iter) : iter_(nullptr) {
+  IteratorWrapperBase() : iter_(nullptr), valid_(false) {}
+  explicit IteratorWrapperBase(InternalIteratorBase<TValue>* _iter)
+      : iter_(nullptr) {
     Set(_iter);
   }
-  ~IteratorWrapper() {}
-  InternalIterator* iter() const { return iter_; }
+  ~IteratorWrapperBase() {}
+  InternalIteratorBase<TValue>* iter() const { return iter_; }
 
   // Set the underlying Iterator to _iter and return
   // previous underlying Iterator.
-  InternalIterator* Set(InternalIterator* _iter) {
-    InternalIterator* old_iter = iter_;
+  InternalIteratorBase<TValue>* Set(InternalIteratorBase<TValue>* _iter) {
+    InternalIteratorBase<TValue>* old_iter = iter_;
 
     iter_ = _iter;
     if (iter_ == nullptr) {
@@ -49,7 +51,7 @@ class IteratorWrapper {
       if (!is_arena_mode) {
         delete iter_;
       } else {
-        iter_->~InternalIterator();
+        iter_->~InternalIteratorBase<TValue>();
       }
     }
   }
@@ -57,7 +59,10 @@ class IteratorWrapper {
   // Iterator interface methods
   bool Valid() const        { return valid_; }
   Slice key() const         { assert(Valid()); return key_; }
-  Slice value() const       { assert(Valid()); return iter_->value(); }
+  TValue value() const {
+    assert(Valid());
+    return iter_->value();
+  }
   // Methods below require iter() != nullptr
   Status status() const     { assert(iter_); return iter_->status(); }
   IteratorSource source() const {
@@ -97,17 +102,12 @@ class IteratorWrapper {
     }
   }
 
-  InternalIterator* iter_;
+  InternalIteratorBase<TValue>* iter_;
   bool valid_;
   Slice key_;
 };
 
-class Arena;
-// Return an empty iterator (yields nothing) allocated from arena.
-extern SourceInternalIterator* NewEmptyInternalIterator(Arena* arena);
+using IteratorWrapper = IteratorWrapperBase<Slice>;
 
-// Return an empty iterator with the specified status, allocated arena.
-extern SourceInternalIterator* NewErrorInternalIterator(const Status& status,
-                                                        Arena* arena);
 
 }  // namespace rocksdb